pyproject.toml
[project]
name = "sre-skills-bench"
version = "0.1.0"
description = "Benchmark suite for evaluating LLMs on Site Reliability Engineering tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "matplotlib>=3.10.7",
]

[project.optional-dependencies]
terraform-generation = [
    "pyyaml>=6.0",
    "boto3>=1.28.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
